Get ViewContactActivity working at a basic level with ContactsProvider2.
With this change the ViewContactActivity is able to display the common
data types defined in ContactsContract. The intents associated with
clicking each item will not necessarily work, as the receiving apps
don't know about the schema changes yet.

Also removed links to ContactsContract.java and SocialContract.java
since those classes have been moved on android.providers and are
accessible without the hard links now.

Made a trivial chane to the fastrack UI by changing the scaleType of the
contact method images from fitCenter to centerInside.