Revert CNAP change to fix more serious Contact List issue
This fix was originally made to handle a case where a CNAP name would
not be logged correctly. However this breaks a more serious scenario
where adding a contact, placing a call to that contact, and then deleting
that contact would still show the name of the old contact in the log,
which is incorrect. Therefore, reverting until a better fix can be found.