Dialer: add onscreen "Dial" button.
Add a new "Dial" button, below the keypad (in portrait) or below the
"digits" widget (in landscape).  It does the exact same action as tapping
the digits widget, and its enabled/disabled state uses the exact same
logic that turns the digit widget green when it's OK to dial.

This is a partial fix for bug 1825144 (soft SEND/END keys in places other
than in-call),

The button has visibility GONE by default; it's enabled on a per-product
basis using a flag in config.xml.  (To enable this on devices that need
it, I'll separately check in resource overlay files in those other
projects)

For now the Dial button is just a regular button, with no special layout
or artwork. (But bug 1993282 is open to cover any future visual design
tweaks we might need.)