Remove redundant ToneGenerator allocation and release.
The code to allocate and release the ToneGenerator object was duplicated
in both onCreate and onResume, and the code to release it was duplicated
in onDestroy and onPause.  We only ever need it while we're in the
foreground, so just do it in onResume / onPause

TESTED: local DTMF tone playback in the dialer.