New dial numbers and backgrounds for the Contacts dialer.
UI overhaul for the dialer in the contacts app.
This CL changes the buttons only.
We lost the voice mail symbol on the '1' button but
the functionality is still there. A separate button
for the VM will be added at the bottom in a separate CL.

Removed the old button assets and added new ones in
the new mdpi and hdpi directories.

Fixed ButtonGridLayout to put the right amount of vertical
padding between buttons.
Also fixed a small bug in the loop to measure each child (was
measuring child 0 instead of child i.

dialpad.xml, converted px to dp units.
Adjusted the size of the buttons and padding.

Bug: 2104523