Preventing pictures from blinking during contact sync.
We were clearing photo cache on every notification from the cursor. There is no reason to do so
and a good reason to not do so. We used to build a sparse array based on the cursor position,
which would have to be invalidated when the cursor re-queried. It's soft cache based on the photo
id, so it can be kept around across those reloads.

Change-Id: I3f7fdcabbbeb09983410f8d092f31647df8fd698