New dial action button at the bottom of the dialer.
There is a new row of 3 buttons (voicemail, dial, backspace).
Currently only the dial button has been implemented.

In the new UI design the dial button should be visible even
on devices with a hard SEND button, therefore I changed the
resource that controls the display to true.

In the dialer, on resume, set the state_enabled attribute properly.

Bug: 2104523