New text field to display the punched numbers.
The numbers are centered in the field.
The active background is used when non empty.
Added top and bottom margin.
The font used appear a bit small compare to the
design, need to come up with a textAppearance resource
to control it (future CL)

Tested on cdma and gsm phones.

Bug: 2104523