Hook up logic for re-parenting of user edits, unit tests.
When persisting edits, we assert that RawContacts.VERSION
has remained consistent.  When this check fails, it usually
means that a server change has changed the underlying data,
and instead of dropping user edits, we "re-parent" their
changes after re-reading the base data.

This change finishes up the re-parenting logic to run over
an entire EntitySet, and also has a nice pile of unit tests
to verify behavior.  Fixes http://b/2115136