Make SDCard/SIM importer allows import into "phone-local" account.
Also make importer try to create "My Contacts" group if it does not exist in
the Google account.

Note that the code for the issue 2138187 is also added but temporaly commented out.
I agree this is not clean change, but please ignore these now.
I'll comment in them soon...

Internal issue number: 2141766