Tweaked the dialer screen to match the designer's PPL.
Created a layout-long-finger version of the dialpad for wvga devices.
For these devices, the buttons are taller than for hvga and
the margin around the elements are also different.

Extracted the bar with the 3 buttons in a separate file that gets
included in both portrait and landscape modes.

Simplified the ButtonGridLayout. Now runs 6% faster (3 runs w/ traceview).
I hardcoded the size of the button grid because the number of columns was
already hardcoded, so why not fix everything?
Use some member to cache the calculation made during onMeasure to be reused
in onLayout.

Tested on dream and sh.... in portrait and landscape modes.

Bug:2104523