Taping the dialed number allows editing.
Previously, taping the dialed number used to dial the number.
Now if non-empty, a cursor shows up at the tap location.
The user can then use the dialpad to delete/insert new numbers.

The previous version of the selectors that drive the number background
has been moved to the non finger drawable because they show the right
background when focused.
In touch mode, the background does not change when focused.

Bug:2160189