Dismiss FastTrack only after calling startActivity().
When launching towards an existing task where the target
activity is already at the top of that task, we triggered
a check in ActivityManager that requires the caller to be
in the foreground.  Before this change, we dismissed the
dialog before calling startActivity(), meaning that in
this case the user action would result in a no-op.

Fixes http://b/2162031