Fix ImportVCardActivity so that NPE will never occur during import.
The NPE occurs only when the importing thread exits before the progress dialog is shown.
In theory (and in practice), this is rare, but anyway, this is a bug.

-----IMPORTANT-----

This change is cherry-picked from the latest build to the eclair.
Please see the description in the issue bellow.

-------------------

Internal issue number: 2164280