Allow TwelveKeyDialer call button to send empty flash for CDMA
Currently the TwelveKeyDialer does not allow the user to press the green
call button if the dial text is empty. However, in CDMA when we place a
3-way call in some network infrastructures, you need to send an empty
Flash command to the network before adding the call. The user is accustomed to
pressing the call button in order to do this. Therefore here we make the
changes in TwelveKeyDialer.java required to handle that keypress and send the
blank flash.

Bug: 2179060
Change-Id: Ib8d05459cac708f9dc216412103c1a324d747ecb