In landscape mode, set the type of digit area to PHONE[do not merge].
Backport from Ieb30191ace8a0d62adec4feaca8f81df7ac4e71b

Previously we blindely set the EditText to NULL to prevent the IME
keyboard to pop up in portrait mode. The issue is that in landscape
mode, the keyboard is not in numlock mode which is a pb on devices
where the 1st line is not made of numbers.

Bug: 2180808