Fix hardcoded px dimensions in "dialpad chooser" UI.
In the "dialpad chooser" UI (which comes up if you launch the Dialer while
a call is already in progress) the 3 items are too tightly spaced on WVGA
devices.  Fixed by changing the relevant dimensions from px to dp.

BUG=2185527
TESTED: got into the "chooser" state on a WVGA phone, confirmed the items
        are now sized correctly.