Fixed the contact's dialer layout.
Currently the bottom row is clipped on WVGA-800.
Changes were made to fix the emergency dialer
that needed to be applied to the contact dialer.

Now, the contact dialer uses the same ButtonGridLayout
class which aligns the button at the bottom of the view.
This way we can control the spacing between the last row and
the 'torpedo' accurately.
The rest of the white space is distributed amongst the element
using layout_weight="1".

Tested on VGA, WVGA-854, WVGA-800

Bug:2198341