EditContactActivity: Restore focus when the screen rotates.
Assign unique but consistent IDs to the EditText on EditContactActivity, so that focus is restored when we recreate the activity.
Also assign IDs to other views, such as GenericEditorView as well.  This will be used to restore views' visibility.

This fixes b/2151340.

Also, set focus at the new EditText when you add a new field by clicking on the "+" button.

This fixes b/2223543.