Extract the full telephone number from the tel: URL.
Previously to extract the number from the dial intent data
we used the Uri class. This does not work very well be cause
dial string can contain character that are special in a Uri context.
For instance '#' in a dial string is a interpreted as a Uri fragment.

To work around that issue, we extract the dial string by flattening
the Uri into a string and skip the 'tel:' scheme.

Bug:2282423