Renamed placeCall to dialButtonPressed.
placeCall does not always place calls so I think the name does not fit.

When we build the call intent, stop adding the number right away, if
number is null, I am not sure what the behavior will be. Instead we add
the tel uri only when we know that the number is not null/empty or when
it is a flash (empty string).

Added branch to insert the last dial number (need one more CL).

In onKeyUp, simplified the logic.

Bug:2227429