Fix the way ExchangeSource determines what order to display names in
In some locales, names should be displayed given name first,
family name second.  In other locales, this is reversed.
All ContactsSources except ExchangeSource base this choice
upon a resource config.  The ExchangeSource was putting family
name first only if the default language was Japanese.
First of all, this is wrong because it should also behave
this way in China and Korea, and secondly, it's just not
clean way to do this when can use a resource instead.

Now the ExchangeSource works the same as all other ContactsSources
in this regard.

Change-Id: I8a640d191b3311487e2aaf92bb7450f72f82afb8