Absorbing extra key events on contact search.
When contact search is started from a keyboard,
we start a search activity, but subsequent key events
are still coming to the original activity and
it just keeps launching new search activities.

Discussed this with Dianne.  The only available solution
in FroYo is to swallow the extra events.

This CL does just that.

Bug: 2565998

Change-Id: Ib9318ad4832e950172af5a3999194bc14321938b