Cleaned up dialButtonPressed.
Made the hard button match the behaviour of the soft button.

In dialButtonPressed:
Moved the comment in the method commen.
Added couple of comments.

Removed the sendflash local parameter, in the case the dial string
is empty we never finish.

Intents are created in block they are used (flash and regular dial)
Added helpers to biuld regular/VM/flash intents.

tested:
On passion:
Last dialed number works even in an add call scenario.
Regular dial works.

On sholes:
Regular dial works.
Last dialed number works when not in call.
Flash is sent when trying to add a call.

Change-Id: I736b108aef90b87f043e68604ffb9b1adbc03f96