Disable dialpad chooser when the user selects "Add call".
As of change https://android-git.corp.google.com/g/59935 , an ACTION_DIAL
intent with no data now brings up the dialpad chooser if the phone is in
use.  This fixed a usability problem with the car home app.

Unfortunately an ACTION_DIAL intent with no data is *also* generated by
the InCallScreen's "Add call" button.  But we obviously *don't* want the
dialpad chooser in that use case, since the whole point of that button is
to add a 2nd call while the phone is already in use.

The fix is to check explicitly for the "Add call" action (using an intent
extra that already exists), and not show the dialpad chooser in that case.

Bug: 2878748
Change-Id: I890cfa72db375ff7b30fc8026da1bc9103118051