Add SipAddress to Contacts UI in master
In gingerbread, change https://android-git.corp.google.com/g/62847 added
SIP addresses to the Contacts app.

This is that same change, manually applied to master.  (I didn't let it
auto-merge because that would have failed: the ViewContactActivity class
is totally gone in master.)

This change is identical to https://android-git.corp.google.com/g/62847
except that the change to ViewContactActivity.java is now applied to
ContactDetailFragment.java.  (The actual code change is identical,
though.)

Bug: 2942998
Change-Id: I18baf54966022d66d387d42d07fe2fdd5c1539f0