Introducing InterpolatingLayout.
Layout similar to LinearLayout. It allows a child
to specify examples of desired size depending on
the parent size. For example if the widget wants to
be 100dip when parent is 200dip and 110dip when
parent is 400dip, the layout will ensure these
requirements and interpolate for other parent sizes.

Change-Id: Id63da83ecab105c46d472cba927a9cead9c2172d