Switching to tab and search UI provided by the framework.
Could not make the search UI always on: apart from some
serious technical difficulties the UX is odd.

First the UX issue: I type in
a query and then switch to some tab.  If we leave the
query string in the field there is a disconnect between
that query and what you see in the list.  If we wipe off
the query and then we have to retype it completely on
the next search.

Now the technical issue: the search field grabs focus
so whenever I open Contacts, the search field is focused
and the keyboard is up.

Also had to hack SearchView a bit because it does not expose
the API of its constituent AutocompleteEditText.

Change-Id: Ie8848d16ab724e9bb6b7025e88b495c7297d49a7