Set ContactInfo.normalizedNumber to null for SIP addresses
At michaelbai's recommendation, set info.normalizedNumber to null after
looking up a SIP address in the contacts database, since SIP addresses
have no concept of a "normalized number".

(This is a followup to https://android-git.corp.google.com/g/71355 , which
was a manual merge from gingerbread.  The manual merge was necessary
because the ContactInfo.normalizedNumber field only exists in master.)

Change-Id: I5b56c0bd132efcc7eba2792cc719a534f289f409