Fix cosmetic glitch in the Call log with SIP addresses
When there's no "label" field to display (like "Home" or "Work"), we were
setting the visibility of that TextView to GONE.  But that's a bad idea,
since we still need it as a RelativeLayout anchor for the "number" field
next to it.  This was causing SIP addresses to move all the way to the
left edge of the screen, where they collide with the "expand" icon.

(I sanity-checked the other RelativeLayout positioning in
recent_calls_list_item_layout.xml, and the only other time we use
visibility=GONE is when that whole bottom row is missing, so we won't have
this same issue there.)

Now the "label" field is INVISIBLE (not GONE) when there's no label to
show.

Also, when there's no label to show we don't need any extra margin
*between* the label and the number.  So in that case, add some negative
left margin to the number view in order to undo the right margin added by
the label view.

TESTED: all combinations of { groups , non-groups } and { real phone
  numbers , SIP addresses } and the label and/or number are now all
  correctly aligned.

Bug: 3062630

Change-Id: I760356ded7b510b08c0791b942507f25ab3a4a4a