Making aggregation suggestion join instant
The old behavior was that you would mark one of the
suggestions for future aggregation, but then
you could still change your mind before saving.

The new behavior is that the second you hit the
"Join" button, the aggregation happens and you stay
on the edit screen.

Also, fixing placement of the aggregation suggestions UI.

Change-Id: I141b996b8b65e7b14711a26dbf66842097338143