Adding the "Edit" mode to aggregation suggestion.
Now if you are creating a new contact and enter
the name of an existing editable contact, the
UI will show an identity card for that other contact
and an Edit button.  Tapping on the Edit button
simply switches the editor to the suggested contact.

The data you may have entered will be transferred to
the suggested contact in another CL.

Change-Id: Ic0af5d883efb17d30dfea052b911e0a5f02d0e91