Preventing multiple instances of contact browser from launching
There is a lot of complexity (and simplifications) in
this CL because as it turns out having fragment transactions
was causing focus to be acquired by the search field,
which was causing all kinds of problems.  So now we only
have one permanent list fragment (which should go into the XML
file) and we just change modes on it.


Bug: 3345023

Change-Id: I096be24b7598350a074eb1dadc954c7619f26669