Fixing handling of single-item data kinds.
The bug was that if we had an existing row
for a single-item kind like nickname,
but that row was empty and therefore invisible,
we would allow the UI to create another,
visible row.

The fix finds that invisible row if it exists
and makes it visible thus eliminating duplication.

Bug: 3334481
Change-Id: Ic639e5ae42b485fe6830078e5043a4f10b757086