Remove tab text in Dialtacts
Remove the text labels from the 4 tabs in the Dialtacts activity, leaving
only the icons.  This allows all 4 tabs to be visible on phone-size
screens in portrait.

This is a short-term fix for ICS dogfooding.  We *will* ultimately bring
back labels at some point, but the exact visual design is still TBD.

Bug: 4520620
Change-Id: I5ecb9a2034bdd4cbc54d298d1318ab6f2d7ed15b