Remove requests to start background thread.
When the activity is being paused or destroyed, we asked the background
thread handling update requests to be stopped. However it is possible
for a request to start the thread is pending. Make sure those are
removed as well.

Since onPause/onDestroy and the request are all handled on the main
thread, there is no race condition between the removal of the request to
start the thread and the call to stopRequestProcessing.

Change-Id: I5eab9fa12d54ce5a5519687946fab8f4c82470b6