Use a cache for the contact info in the call log.
This commit introduces an ExpirableCache, a cache whose elements can be
marked as expired.

Instead of clearing the cache when the activity is resume, marked the
items in the cache as expired. Items that are found in the cache but are
expired will trigger a re-fetch of the contact details.

Change-Id: Ib06534c33d3427047afc4d95fc254a644a92d836