Refactor the code to determine the details of a call.
This same logic is (partly) duplicated in the call log for formatting
the individual calls in the list. We need to get these two to be the
same so what is best than having a single copy of the code?

This tries to pull as much as possible of the logic into the views class
and leave the rest of the code to handle only extracting the values from
the source.

Change-Id: Ic42b4de70b0a8cb6101f80504d247b68e138a44c