Contacts App modifications for large photos.
Most of the work is on the provider side - essentially, the
provider will now deal with any size photo, downsizing it as
necessary to fit in a display size and thumbnail size.  So
on the client we just need to pass in a larger photo to get a
hi-res display image for the contact.

Also switched a few spots to use the PHOTO_URI (with
openAssetFile) instead of PHOTO_THUMB_URI - more likely
remain.

Change-Id: I6ff8dd5453a9ff787907ab113d6e8aa6a71fcae1