Uses a tag for vCard-related notifications.
Currently, we are adding notifications for vCard events using an
identifier. However, these identifiers need to be unique within the
application, which is not easy in this case since an unbounded number of
identifier can be used (i.e., they are based on the job id of the vCard
being imported).

Instead, add a tag to the notification, so that notifications for vCard
cannot conflict (and replace) notifications coming from somewhere else
in the application.

Change-Id: Ia544fc4e11c43799222c30cece9330478208fd46