Contact card on tablet
- Two-column layout in landscape
- ViewPager and tab carousel in portrait
- Views with and without social updates
- Use ContactDetailLayoutController which controls hiding/showing
fragments based on orietnation and contact data
- Move options menu into the loader fragment because there can
be multiple ContactDetailFragments on the screen now at one time
or it can be dynamically created by the ViewPagerAdapter and
we won't have a reference to that in the PeopleActivity

TODO: Use ContactDetailLayoutController in ContactDetailActivity
so there isn't duplicated code. But just make it work now and do
cleanup later.

Change-Id: Ibf8eb317c6e19bc3d98eb9ce46a45cccb237c631