Group list: Don't reset scroll position during sync
Manually call ListView.onSave/RestoreInstanceState to restore the scroll
position after setting a new adapter.

Note this solution isn't perfect; ListView.onSaveInstanceState saves
the item ID of the fist item (as well as offset to the scroll, etc).
The scroll position will only be restored correctly if the ID is valid
(i.e. non negative) and unique.

In our case, if the first item is a header, the list view will get
confused and may scroll to the wrong position.  But it shouldn't be that
big deal here...

Change-Id: I07e2682178d068253ccc96bc7a836de4bd25c7ef