Fetch contact details as soon as possible.
If we currently do not have contact details available, do not wait one
second since the first redraw to start the thread that fetches the
contact details. Instead, start the thread right away, do reduce the
latency of displaying the contact details.

If, however, we simply want to refresh the content of the existing
cache, keep the delay as before: in that case, there is no particular
hurry.

Change-Id: Iecef7a47f972641bed53d47b7b22dd115410583a