Add action to copy the phone number in the dialer.
This is different from the call action as it does not actually dial the
number but allows the user to edit the number before dialing.

As for the delete action, this might not be its final location, but this
moves it to the right activity, the call details page instead of the
long-press menu in the call log.

Change-Id: I0110f7f876d3ee34adc0461a858a8ab1e694f1bd