Implementation of business logic for voicemail status.
The new class VoicemailStatusHelper implements all the business logic to
determine what status messages needs to be shown to the user based on
the current status of various voicemail sources. It also orders the
messages in the order of priority.

A follow up cl would use this class to show the message and action to
the user.

Change-Id: I1d0321598356d7451715f027d00ceb3c09418669