If there are no new calls, do not show older header.
If there are new and older calls, we should headers to identify them.
However, if there are only older calls, it does not make sense to have a
header there, and we should just show the call themselves.

On the other hand, if there are only new calls, it still makes sense to
keep the header there.

Bug: 5035679
Change-Id: I717577cf3b76f95843050acad4511ef71c864123