Fix bug 5036399 (2-pane) Opening a contact from quick
... contact won't update current view.

The issue is that when quick contact opens the activity's
onSaveInstanceState() gets called, where we clear the action bar adapter
listener.
Then, when the user selects "contacts" on quick contact, onNewIntent()
gets called and we update the current tab there.
However, because it's onResume() where we restore the listener and onResume()
is called after onNewIntent(), the activity misses the current tab change
callback.

Fix is easy -- just always call updateFragmentsVisibility() in onResume()
when we restore the listener to make sure fragments visibility matchs
the current tab.

(updateFragmentsVisibility() is virtually no-op if there's no changes, so
always calling it is not inefficient.)

Change-Id: I64aa229cbb1aa37114fce73e83849fbb2098acc7