Handle call log entries with a NULL value for the NEW field.
The current query causes SQLite to not return items with a NEW field set
to NULL (if they are missed calls or voicemail) neither in the NEW
query nor its negation.

This fix makes sure that each entry will match one of the two possible
queries and therefore always appear somewhere in the call log.

Bug: 5036125
Change-Id: I66e0b9ed50c9d5733e14ec8da181884a0a6f160d