Fix for loading the same contact photo twice.
Currently, the code in the ContactPhotoManager which loads the pictures
from the database assumes that each picture is requested only once.
However, as in the case of the call log, the same picture might be
needed more than once.

This commit changes the code to create a set of photos to load instead
of a list, so that each id or URI appears in the least at most once.
This fixes the problem with the previous code which would end up
resetting each picture loaded more than once in the same request to the
default contact picture just after loading it.

Bug: 5020196
Change-Id: I82169e9cf9048a31e2382fbfa962df5cfdd0971f