Made voicemail status query from calllogfragment asynchronous.
callLogQueryHandler is now used to fetch voicemail_status cursor and
pass that to VoicemailStatusHelper to compute the messages for display.
VoicemailStatusHelper no longer make direct call to the content
provider, so is now safe to be used from the ui thread.

Also, modified status message layout to have visibility=gone as a
default value.

Change-Id: I2dce71080d974c318adcbd134e91df85ac6ce611