Show all contacts when search query is empty.
DefaultContactListAdapter seems to be used by the contact picker too,
which looks to be still working fine with this CL.

This will arguably make worse the search + orientation change issue; with this
CL if you rotate the device while searching, the initial list after rotation
always shows all contacts.
I'll fix this in a different CL, as that's not really a new bug.
For now I just filed bug 5042558.

Bug 4983521

Change-Id: I259086511717625cf5925d6ab45b16b62d742dbd