Update the content of the call log when resumed.
Since the call log is now a fragment, onResume is not sufficient to
determine whether the user came back to the call log.

Instead use both onResume and onVisibilityChanged. However,
onVisibilityChanged can be called before onCreate and other lifecycle
methods are, therefore we cannot assume the fragment is in proper state.

As a workaround, use the isResumed() method to determine whether an
update should happen.

Change-Id: I47cdf07278447209efa0349c7d0a2943e206bdb5