Correctly update contact information when resumed.
Currently the contact information was not resumed because the content of
the cache was present and the application assumed that we had a
duplicate request.

However, since the only thing we need for a request is the phone number,
use a list of strings and do not add to the list if there is already the
same value in it. This prevents duplicate requests.

Also remove CallerInfoQuery since it is no longer needed.

Bug: 4579309
Change-Id: Ife971dfdd8c9e554689bd00d3be49aec667e43f6