Dismiss missed calls when the call log is shown.
The notification used to be dismissed by the CallLogActivity. Now that
that class is no longer used but the CallLogFragment is used instead,
add the dismissal in the fragment.

With the fragment, detecting when the code is being shown is a bit more
complex (it takes into account both onResume and onVisibilityChanged).

Bug: 4521535
Change-Id: I01af5b3ab0e1e66e9b90d93b4957c0dff9816603